requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
pymongo==4.15.4
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.3.0
pydantic==2.12.4
httpx==0.27.0
